Change .local -> .home
It is a bad idea to use *.local* domain name because it's use by
mDNS or zeroconf. So it can make implicit conflict between
machine hostname and ingress path. If a match occure between
hostname and ingress name, i do not know out it is resolved and I
want to avoid any confusion.